+++ ChangeLog.md
@@ -4,6 +4,11 @@
 
 ## Upcoming Release
 
+Blob:
+
+- Fixed get service properties or account properties failure from blob endpoint when Uri has suffix '/' after account name.
+- Fixed get system container failure.
+
 ## 2021.6 Version 3.13.0
 
 General: